將Sidebar獨立出來，並新增去背Logo
