+++ src/Components/Sidebar/Sidebar.css
@@ -1,0 +1,171 @@
+@import url("https://fonts.googleapis.com/css2?family=Lato&display=swap");
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  /* font-family: Simsun, "Lato", sans-serif; */
+  font-family: "Lato", sans-serif;
+}
+
+.collect,
+.notify,
+.setting,
+.draft {
+  display: flex;
+  height: 90vh;
+  align-items: center;
+  justify-content: center;
+  font-size: 3rem;
+  background-color: #000;
+  color: #fff;
+}
+
+.navbar {
+  background-color: #ffffff;
+  height: 80px;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.menu-bars {
+  margin-left: 1rem;
+  padding-right: 15px;
+  font-size: 2rem;
+  background: none;
+  color: white;
+}
+
+.line{
+  width: 100%;
+  margin-bottom: 4px;
+  margin-top: 4px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-bottom: 1.5px solid lightgray;
+}
+
+.nav-menu {
+  background-color: #f3f3f3;
+  border-right: 1px solid #1e1e1e;
+  width: 210px;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  left: -250px;
+  transition: 200ms;
+}
+
+.nav-menu.active {
+  left: 0;
+  transition: 200ms;
+}
+
+
+.nav-text {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  padding: 8px 0px 8px 16px;
+  font-weight: bold;
+  list-style: none;
+  height: 60px;
+}
+
+.nav-text a {
+  text-decoration: none;
+  color: #343434;
+  font-size: 18px;
+  width: 90%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  padding: 0 16px;
+}
+
+.nav-text a:hover {
+  background-color: #000000;
+  color: #fff;
+  border-radius: 10px;
+}
+
+.nav-menu-items {
+  width: 100%;
+}
+
+.search__container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 12px;
+  margin-top: 12px;
+}
+  
+.searchBar {
+  padding: 12px 18px;
+  outline: 0;
+  border: none;
+  box-shadow: 4px 4px 10px rgba(163,177,198,0.5), -4px -4px 10px rgba(255, 255, 255, 0.7);
+  background-image: url("data:image/svg+xml;charset=utf8,%3Csvg  xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24'%3E%3Cpath d='M15.5 14h-.79l-.28-.27C15.41 12.59 16 11.11 16 9.5 16 5.91 13.09 3 9.5 3S3 5.91 3 9.5 5.91 16 9.5 16c1.61 0 3.09-.59 4.23-1.57l.27.28v.79l5 4.99L20.49 19l-4.99-5zm-6 0C7.01 14 5 11.99 5 9.5S7.01 5 9.5 5 14 7.01 14 9.5 11.99 14 9.5 14z'/%3E%3Cpath d='M0 0h24v24H0z' fill='none'/%3E%3C/svg%3E");
+  background-repeat: no-repeat;
+  background-size: 18px 18px;
+  background-position: center right 1rem;
+  background-color: #f3f3f3;
+  border-radius: 20px;
+  transition: padding 100ms ease-in-out;
+}
+
+.searchBar:hover {
+	padding: 15px 22.5px;
+}
+
+.searchBar:focus {
+	padding: 15px 22.5px;
+}
+
+.navbar-title {
+  background-color: #f3f3f3;
+  width: 100%;
+  height: 50px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-weight: bold;
+  letter-spacing: 3px;
+  font-size: 22px;
+  margin-top: 16px;
+}
+
+span {
+  margin-left: 16px;
+}
+
+.bottomBlock{
+  position:absolute;
+  bottom: 0; 
+  width: 100%; 
+  height: 85px; 
+  /* border: 1px; */
+  background-color: #f3f3f3;
+}
+
+.logo{
+  display: flex;
+  width: 90%;
+  margin-left: 0.5em;
+  /* padding-top:4px; */
+  /* padding-bottom: 20px; */
+}
+
+.main__item{
+  height: calc(100vh - 250px);
+  bottom: 0;
+  z-index: 500;
+  display: block;
+  overflow-y: auto; /* Scrollable contents if viewport is shorter than content. */
+  border-bottom: 0
+}
